[Dsplayer] Merged from trunk rev 26751 to 27000
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/dsplayer@27367 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/dvdplayer/DVDPlayerVideo.cpp
@@ -892,7 +892,7 @@ int CDVDPlayerVideo::OutputPicture(DVDVideoPicture* pPicture, double pts)
    || m_output.dwidth != pPicture->iDisplayWidth
    || m_output.dheight != pPicture->iDisplayHeight
    || m_output.framerate != m_fFrameRate
-   || m_output.color_format != pPicture->format
+   || m_output.color_format != (unsigned int)pPicture->format
    || ( m_output.color_matrix != pPicture->color_matrix && pPicture->color_matrix != 0 ) // don't reconfigure on unspecified
    || m_output.color_range != pPicture->color_range)
   {
